# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finalechat API",
    "version": "1.0.0",
    "summary": "The app where agents message the user.",
    "description": "Threads of messages and questions between agent sessions and one user, with long-polling, server-sent events and Web Push. Human-readable reference: https://www.finalechat.com/api/ . Agent quick start: https://www.finalechat.com/AGENTS.md .",
    "contact": {
      "url": "https://www.finalechat.com"
    }
  },
  "servers": [
    {
      "url": "https://www.finalechat.com/api/v1"
    },
    {
      "url": "https://api.finalechat.com/api/v1"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/threads/{ref}/attachments": {
      "post": {
        "tags": [
          "Attachments"
        ],
        "summary": "Upload files to attach to a later message",
        "description": "Creates a missing ext: thread on demand. Send multipart/form-data with one or more file parts (any field name; at most 8), or a raw body whose Content-Type is the file type with the filename in X-Filename or ?filename=. Images (png, jpeg, gif, webp) get dimensions and a JPEG thumbnail. Uploads not attached to a message within 24 hours are deleted.",
        "parameters": [
          {
            "name": "ref",
            "in": "path",
            "required": true,
            "description": "Thread UUID or `ext:<external_id>`.",
            "schema": {
              "type": "string"
            },
            "examples": {
              "uuid": {
                "value": "01a07a60-6dab-780a-bf4d-20ef15c0d7d7"
              },
              "external": {
                "value": "ext:claude-code:7f3a9c2e"
              }
            }
          },
          {
            "name": "filename",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filename for a raw-body upload (alternative to X-Filename)."
          },
          {
            "name": "X-Filename",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Filename for a raw-body upload."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    },
                    "maxItems": 8
                  }
                }
              }
            },
            "*/*": {
              "schema": {
                "type": "string",
                "format": "binary"
              },
              "example": "raw file bytes; Content-Type names the file type"
            }
          }
        },
        "responses": {
          "201": {
            "description": "The pending attachments.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "attachments",
                    "thread_id"
                  ],
                  "properties": {
                    "attachments": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Attachment"
                      }
                    },
                    "thread_id": {
                      "type": "string",
                      "format": "uuid",
                      "description": "UUID."
                    }
                  }
                },
                "example": {
                  "attachments": [
                    {
                      "id": "01a07ab9-7a53-7da4-81ad-3f7bad1f53db",
                      "thread_id": "01a07ab9-312a-776f-ae4b-48c3cba6530c",
                      "message_id": null,
                      "kind": "file",
                      "content_type": "text/plain",
                      "filename": "notes.txt",
                      "size": 23,
                      "created_at": "2026-09-07T07:16:10.963984Z",
                      "url": "/api/v1/attachments/01a07ab9-7a53-7da4-81ad-3f7bad1f53db"
                    }
                  ],
                  "thread_id": "01a07ab9-312a-776f-ae4b-48c3cba6530c"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "A file is over 10 MiB or the request is too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Empty file, undecodable image, or too many files",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Object storage failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Attachments are not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/attachments/{id}": {
      "get": {
        "tags": [
          "Attachments"
        ],
        "summary": "Download an attachment",
        "description": "Streams the original bytes with the stored Content-Type. Bearer or session auth.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID."
            },
            "description": "Attachment id."
          }
        ],
        "responses": {
          "200": {
            "description": "The file.",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "inline for images, PDFs and text; attachment otherwise; carries the filename."
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "private, max-age=31536000, immutable"
              }
            },
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Object storage failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Attachments are not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "head": {
        "tags": [
          "Attachments"
        ],
        "summary": "Attachment headers only",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID."
            },
            "description": "Attachment id."
          }
        ],
        "responses": {
          "200": {
            "description": "Headers as for GET, no body."
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/attachments/{id}/thumb": {
      "get": {
        "tags": [
          "Attachments"
        ],
        "summary": "Download an image's thumbnail",
        "description": "JPEG thumbnail with the longest side at most 640 pixels. For a non-image attachment the original bytes are returned.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "UUID."
            },
            "description": "Attachment id."
          }
        ],
        "responses": {
          "200": {
            "description": "The thumbnail.",
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "inline for images, PDFs and text; attachment otherwise; carries the filename."
              },
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "private, max-age=31536000, immutable"
              }
            },
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Object storage failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Attachments are not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "fc_ token",
        "description": "API token created in the app under Settings → Agents. Usually supplied as the FINALECHAT_TOKEN environment variable."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fc_session",
        "description": "Browser session set by POST /auth/login. State-changing requests must be same-origin."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "bad_request",
                  "unauthorized",
                  "invalid_token",
                  "invalid_credentials",
                  "forbidden",
                  "csrf",
                  "signup_closed",
                  "invalid_invite",
                  "not_found",
                  "conflict",
                  "invalid_state",
                  "already_resolved",
                  "email_taken",
                  "no_subscriptions",
                  "too_large",
                  "validation_failed",
                  "rate_limited",
                  "push_disabled",
                  "internal_error",
                  "method_not_allowed",
                  "storage_unavailable",
                  "attachments_disabled"
                ]
              },
              "message": {
                "type": "string"
              }
            }
          }
        },
        "example": {
          "error": {
            "code": "validation_failed",
            "message": "prompt is required."
          }
        }
      },
...
      "Attachment": {
        "type": "object",
        "required": [
          "id",
          "thread_id",
          "message_id",
          "kind",
          "content_type",
          "filename",
          "size",
          "created_at",
          "url"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID."
          },
          "thread_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID."
          },
          "message_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Null until the upload is attached to a message."
          },
          "kind": {
            "type": "string",
            "enum": [
              "image",
              "file"
            ],
            "description": "image for png, jpeg, gif and webp (sniffed); file otherwise."
          },
          "content_type": {
            "type": "string"
          },
          "filename": {
            "type": "string",
            "description": "Sanitised original filename, at most 200 characters."
          },
          "size": {
            "type": "integer",
            "description": "Bytes."
          },
          "width": {
            "type": "integer",
            "description": "Images only."
          },
          "height": {
            "type": "integer",
            "description": "Images only."
          },
          "thumb_width": {
            "type": "integer",
            "description": "Images only; longest side at most 640."
          },
          "thumb_height": {
            "type": "integer",
            "description": "Images only."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "UTC RFC 3339 timestamp."
          },
          "url": {
            "type": "string",
            "description": "Download path relative to the base URL, e.g. /api/v1/attachments/{id}."
          },
          "thumb_url": {
            "type": "string",
            "description": "JPEG thumbnail path; images only."
          }
        },
        "example": {
          "id": "01a07ab9-3132-7b58-85f0-5c74ef773395",
          "thread_id": "01a07ab9-312a-776f-ae4b-48c3cba6530c",
          "message_id": "01a07ab9-314c-72db-9505-30a34c350400",
          "kind": "image",
          "content_type": "image/png",
          "filename": "shot.png",
          "size": 168591,
          "width": 900,
          "height": 300,
          "thumb_width": 640,
          "thumb_height": 213,
          "created_at": "2026-09-07T07:15:52.264313Z",
          "url": "/api/v1/attachments/01a07ab9-3132-7b58-85f0-5c74ef773395",
          "thumb_url": "/api/v1/attachments/01a07ab9-3132-7b58-85f0-5c74ef773395/thumb"
        }
      },
...
    }
  }
}